Setup for better asset adding implementation
Currently, tracked does not take the user to a new editable AssetPage. But
instead, creates 'empty' asset which then the user needs to click on the newly
created asset and edit then save. In the future this process will be quicker in
a way such that when the user scans an asset, they will be directly taken to
the asset page with scanned field already filled (doe or s/n) out just waiting
for them to hit save.